Creacion de test unitarios en dominio e insfraestructura

+++ microservicio/dominio/src/main/java/com/ceiba/pedido/puerto/dao/DaoPedido.java
@@ -8,5 +8,5 @@ public interface DaoPedido {
 
     List<DtoPedido> listar();
 
-    Long contarPedidosEntregados(Long idCliente);
+    Integer contarPedidosEntregados(Long idCliente);
 }
